requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-elevenlabs
livekit-plugins-noise-cancellation
livekit-plugins-turn-detector
livekit-plugins-groq
livekit-plugins-cartesia
livekit-plugins-deepgram
mem0ai
duckduckgo-search
langchain_community
ics
pytz
torch
requests
python-dotenv